requirements.txt
fastapi
uvicorn
streamlit
requests
nltk
sentence-transformers
newspaper3k
transformers
torch
